requirements.txt
# 라즈베리파이5 전용 라이브러리 리스트

# absl-py==2.1.0
# astunparse==1.6.3
# blinker==1.9.0
# certifi==2025.1.31
# charset-normalizer==3.4.1
# click==8.1.8
# Flask==3.1.0
# flatbuffers==25.9.23
# gast==0.6.0
# google-pasta==0.2.0
# grpcio==1.70.0
# h5py==3.12.1
# idna==3.10
# itsdangerous==2.2.0
# Jinja2==3.1.5
# keras==3.8.0
# libclang==18.1.1
# Markdown==3.7
# markdown-it-py==3.0.0
# MarkupSafe==3.0.2
# mdurl==0.1.2
# ml-dtypes==0.4.1
# namex==0.0.8
# numpy==2.0.2
# opencv-python==4.11.0.86
# opt_einsum==3.4.0
# optree==0.14.0
# packaging==24.2
# protobuf==5.29.3
# Pygments==2.19.1
# pyserial==3.5
# requests==2.32.3
# rich==13.9.4
# six==1.17.0
# tensorboard==2.18.0
# tensorboard-data-server==0.7.2
# tensorflow==2.18.0
# tensorflow-io-gcs-filesystem==0.37.1
# termcolor==2.5.0
# typing_extensions==4.12.2
# urllib3==2.3.0
# Werkzeug==3.1.3
# wrapt==1.17.2

# macos전용(m1) 라이브러리 리스트 

# TensorFlow on Apple Silicon (Metal 가속)
tensorflow-macos==2.12.0
tensorflow-metal==1.0.0

# TF 2.12의 Numpy 제약(NumPy 2.x 미지원) 반영
numpy==1.23.5
h5py==3.9.0
grpcio==1.54.3
protobuf==3.20.3

# CV/유틸
opencv-python==4.10.0.84
absl-py==2.1.0
astunparse==1.6.3
google-pasta==0.2.0
opt_einsum==3.4.0
six==1.17.0
typing_extensions==4.12.2
packaging==24.2
pyserial==3.5
requests==2.32.3